# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Deel home task API"
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": ["http"],
  "paths": {
...
    "/jobs/unpaid": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get all unpaid jobs for a user (either a client or contractor), for active contracts only",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "profile_id",
            "type": "number",
            "in": "header",
            "description": "Id of user profile",
            "required": true,
            "default": 2
          }
        ],
        "responses": {
          "200": {
            "description": "successful"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/jobs/{job_id}/pay": {
      "post": {
        "tags": ["jobs"],
        "summary": "Pay for a job, a client can only pay if his balance >= the amount to pay. The amount should be moved from the client's balance to the contractor balance",
        "description": "notice that you can pay for a job only once, if you try to do it twice with the same params, it will fail",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "profile_id",
            "type": "number",
            "in": "header",
            "description": "Id of user profile",
            "required": true,
            "default": 1
          },
          {
            "name": "job_id",
            "in": "path",
            "description": "job id",
            "required": true,
            "type": "number",
            "default": 1
          }
        ],
        "responses": {
          "200": {
            "description": "successful"
          },
          "400": {
            "description": "bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "internal server error"
          }
        }
      }
    },
...
  }
}